// src/app/models/usuarios.interface.ts
export interface UsuariosInterface {
    //Atributos del Usuario
    ID_Usuario?: string;
    Rut_Usuario?: string;
    Nombres_Usuario?: string;
    Apellido_Paterno_Usuario?: string;
    Apellido_Materno_Usuario?: string;
    Fecha_Nacimiento_Usuario?: Date;
    Sexo_Usuario?: string;
    Direccion_Usuario?: string;
    Email_Usuario?: string;
    Telefono_Usuario?: string;
   }
   